requirements.txt
pillow
requests
sentence-transformers
faiss-cpu
numpy
opencv-python
pytesseract
